Changed how GazeManager is initialized.
It is now always activated when the program is running. This is because of how the API works. It will also let me rely on the API for various information, instead of Java, e.g. get screen sizes from the API itself, instead of GraphicsEnvironment.

When capture starts and stops, the listener is now added and removed, instead of the manager being activated and deactivated.